+++ packages/taiko-client/prover/proof_producer/zkvm_producer.go
@@ -76,9 +76,9 @@ func (s *ZKvmProofProducer) RequestProof(
 		"Request zk proof from raiko-host service",
 		"blockID", blockID,
 		"coinbase", meta.GetCoinbase(),
-		"height", header.Number,
 		"hash", header.Hash(),
-		"zk type", s.ZKProofType,
+		"zkType", s.ZKProofType,
+		"time", time.Since(requestAt),
 	)
 
 	if s.Dummy {
@@ -123,6 +123,10 @@ func (s *ZKvmProofProducer) Aggregate(
 	log.Info(
 		"Aggregate zkvm batch proofs from raiko-host service",
 		"zkType", s.ZKProofType,
+		"batchSize", len(items),
+		"firstID", items[0].BlockID,
+		"lastID", items[len(items)-1].BlockID,
+		"time", time.Since(requestAt),
 	)
 	if len(items) == 0 {
 		return nil, ErrInvalidLength
@@ -173,7 +177,7 @@ func (s *ZKvmProofProducer) callProverDaemon(
 
 	output, err := s.requestProof(zkCtx, opts)
 	if err != nil {
-		log.Error("Failed to request proof", "height", opts.BlockID, "error", err, "endpoint", s.RaikoHostEndpoint)
+		log.Error("Failed to request proof", "blockID", opts.BlockID, "error", err, "endpoint", s.RaikoHostEndpoint)
 		return nil, err
 	}
 
@@ -192,10 +196,15 @@ func (s *ZKvmProofProducer) callProverDaemon(
 	}
 	log.Info(
 		"Proof generated",
-		"height", opts.BlockID,
+		"blockID", opts.BlockID,
 		"time", time.Since(requestAt),
 		"producer", "ZKvmProofProducer",
 	)
+	if s.ZKProofType == ZKProofTypeR0 {
+		metrics.ProverR0ProofGenerationTime.Set(float64(time.Since(requestAt).Seconds()))
+	} else if s.ZKProofType == ZKProofTypeSP1 {
+		metrics.ProverSP1ProofGenerationTime.Set(float64(time.Since(requestAt).Seconds()))
+	}
 
 	return proof, nil
 }
@@ -509,6 +518,12 @@ func (s *ZKvmProofProducer) requestBatchProof(
 		"time", time.Since(requestAt),
 		"producer", "ZKvmProofProducer",
 	)
+
+	if s.ZKProofType == ZKProofTypeR0 {
+		metrics.ProverR0AggregationGenerationTime.Set(float64(time.Since(requestAt).Seconds()))
+	} else if s.ZKProofType == ZKProofTypeSP1 {
+		metrics.ProverSP1AggregationGenerationTime.Set(float64(time.Since(requestAt).Seconds()))
+	}
 
 	return proof, nil
 }